+++ lib/src/domain/services/i_local_auth_service.dart
@@ -2,7 +2,13 @@ import '../../domain/enums/biometric_type.dart';
 import '../../domain/interfaces/either.dart';
 
 abstract class ILocalAuthService {
-  Future<Either<Exception, bool>> authenticate({String localizedReason = ''});
+  Future<Either<Exception, bool>> authenticate({
+    String localizedReason = '',
+    bool useErrorDialogs = true,
+    bool stickyAuth = false,
+    bool sensitiveTransaction = true,
+    bool biometricOnly = false,
+  });
   Future<Either<Exception, bool>> canCheckBiometrics();
   Future<Either<Exception, bool>> isDeviceSupported();
   Future<Either<Exception, Unit>> stopAuthentication();